reccon-system-backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.59
botocore==1.42.59
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.12.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
